+++ index.html
@@ -95,6 +95,6 @@
         </div>
       </div>
     </div>
-    <script src="index.js"></script>
+    <script type="module" src="index.js"></script>
   </body>
 </html>
