+++ modules/BookClass.js
@@ -1,0 +1,9 @@
+class Book {
+  constructor(title, author, id) {
+    this.id = id;
+    this.title = title;
+    this.author = author;
+  }
+}
+
+export default Book;
