+++ package.json
@@ -16,5 +16,16 @@
   "bugs": {
     "url": "https://github.com/paulinagonzalezc/awesome-books-es6/issues"
   },
-  "homepage": "https://github.com/paulinagonzalezc/awesome-books-es6#readme"
+  "homepage": "https://github.com/paulinagonzalezc/awesome-books-es6#readme",
+  "devDependencies": {
+    "babel-eslint": "^10.1.0",
+    "eslint": "^7.32.0",
+    "eslint-config-airbnb-base": "^14.2.1",
+    "eslint-plugin-import": "^2.26.0",
+    "hint": "^7.1.3",
+    "stylelint": "^13.13.1",
+    "stylelint-config-standard": "^21.0.0",
+    "stylelint-csstree-validator": "^1.9.0",
+    "stylelint-scss": "^3.21.0"
+  }
 }
